requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
streamlit==1.31.0
python-multipart==0.0.6

# LLM and Embeddings
google-generativeai==0.3.2
sentence-transformers==2.3.1

# Vector DB
faiss-cpu==1.7.4

# Data processing
pandas==2.2.0
numpy==1.26.3
openpyxl==3.1.2

# PDF processing
PyPDF2==3.0.1
pdfplumber==0.10.3

# Utilities
python-dotenv==1.0.1
tiktoken==0.5.2
pydantic==2.6.0
